requirements.txt
selenium==4.27.1
pytest==8.3.4
allure-pytest==2.13.5
webdriver-manager==4.0.2
